+++ datacrunch_compat/datacrunch/InferenceClient/__init__.py
@@ -1,3 +1,3 @@
-from verda.InferenceClient import InferenceClient, InferenceResponse
+from verda.inference_client import InferenceClient, InferenceResponse
 
 __all__ = ['InferenceClient', 'InferenceResponse']