+++ datacrunch_compat/datacrunch/InferenceClient/inference_client.py
@@ -1,1 +1,1 @@
-from verda.InferenceClient.inference_client import *
+from verda.inference_client import *