+++ datacrunch_compat/datacrunch/datacrunch.py
@@ -2,19 +2,19 @@
 
 from verda import VerdaClient as DataCrunchClient
 from verda._version import __version__
-from verda.authentication.authentication import AuthenticationService
-from verda.balance.balance import BalanceService
+from verda.authentication import AuthenticationService
+from verda.balance import BalanceService
 from verda.constants import Constants
-from verda.containers.containers import ContainersService
-from verda.http_client.http_client import HTTPClient
-from verda.images.images import ImagesService
-from verda.instance_types.instance_types import InstanceTypesService
-from verda.instances.instances import InstancesService
-from verda.locations.locations import LocationsService
-from verda.ssh_keys.ssh_keys import SSHKeysService
-from verda.startup_scripts.startup_scripts import StartupScriptsService
-from verda.volume_types.volume_types import VolumeTypesService
-from verda.volumes.volumes import VolumesService
+from verda.containers import ContainersService
+from verda.http_client import HTTPClient
+from verda.images import ImagesService
+from verda.instance_types import InstanceTypesService
+from verda.instances import InstancesService
+from verda.locations import LocationsService
+from verda.ssh_keys import SSHKeysService
+from verda.startup_scripts import StartupScriptsService
+from verda.volume_types import VolumeTypesService
+from verda.volumes import VolumesService
 
 # for `from datacrunch.datacrunch import *`
 __all__ = [