+++ tests/unit_tests/ssh_keys/test_ssh_keys.py
@@ -2,7 +2,7 @@ import pytest
 import responses  # https://github.com/getsentry/responses
 
 from verda.exceptions import APIException
-from verda.ssh_keys.ssh_keys import SSHKey, SSHKeysService
+from verda.ssh_keys import SSHKey, SSHKeysService
 
 INVALID_REQUEST = 'invalid_request'
 INVALID_REQUEST_MESSAGE = 'Your existence is invalid'