+++ verda/InferenceClient/__init__.py
@@ -1,3 +1,0 @@
-from .inference_client import InferenceClient, InferenceResponse
-
-__all__ = ['InferenceClient', 'InferenceResponse']